moved duplicate get data code to its own file
